refactor: Activity Model 코드 리팩토링

+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/member/Member.kt
@@ -2,5 +2,5 @@ package com.emmsale.data.member
 
 data class Member(
     val name: String,
-    val activityIds: List<Int>,
+    val activityIds: List<Long>,
 )
